Convert `version-check` task to gulp
